lazy load latest package definitions with yumhelper
I could observe that yumhelper.py can run quite some time
and use quite a lot of memory (up to 100MB!) if you're
using many repositories. As the yumhelper is only needed
if you use latest as a ensure-keyword it makes no sense
for people like me letting puppet fetch these details
always, if you're never going to use latest. Therefor this
patch changes the behaviour, that the yumhelper is only
called, if @updates is accessed the first time.
=> lazy loading.