lazy load latest package definitions with yumhelper 2.2
I have to admit that the version 2.1 one of this patch
was not good tested from my side and it simply didn't work.
I'm very sorry for that.

However I did now extensive (manual) testing and it works
the way it should for me.
I can confirm that yumhelper isn't called if there is no
latest ensure and it is called if there is at least one.

So correct commit message:

After discussion on the list, this is the new version for
lazy loading yum latest package versions. It is implemented
mainly in the way that Luke proposed. However it stores
the latest informations in the variable latest_info so
the latest method didn't get too hackish and could nearly be
left like it was before.
+ fixed copy&paste regression, so :latest_info is not anymore
on the class side
+ fixed package iteration and ensure usage.