Propose fix/workaround for wrong facter variables in templates
Due to the fix for #1427 variables are now real instance
variables. Therfore we have to escape special characters
like ':' and '.' for variable names. Otherwise templates and with
that puppet will fail.
This patch will introduce an escaping (_) for such
variables.
It is clear that things like that should be fixed in facts,
however if one fact will contain a wrong character, the whole
puppet run will fail. Applying this patch will prevent puppet
to fail due to wrong characters in facter variable names.