enhance redhat puppetmaster init.d script to easy start puppetmaster as a mongrel cluster v2.1
If you'd like to run puppetmaster in a mongrel cluster you have to start X

instances of puppetmaster with servertype set to mongrel. With this patch
you can add more than one port to the sysconfig file and puppetmaster will

automatically be started with severtype set to mongrel.
You can now even specify an alternate port in the sysconfig file on which
puppetmaster should be run as webrick based server. See sysconfig file for

more documentation.

This is the second version of this patch and it adresses some comments
which have been made by David Lutterkort on Ticket #1460:

- there's no need to introduce PUPPETMASTER_PID_DIR as a variable in
  sysconfig

removed.

- the way RETVAL is accumulated (RETVAL=$? || $RETVAL) is wrong;
  it needs to be 'ret=$?; [ $ret != 0 ] && RETVAL=$ret'

fixed.

- make sure that 'status' still works and lists all the instances in
  the multiport case (similar to 'service nfs status')

added.

v2.1:

addressed changing [ -n "$PUPPETMASTER_PORTS" ] && [ ${#PUPPETMASTER_PORTS[@]} -gt 1 ] to [ $multiport = yes