enable maillist on centos, redhat, fedora
Fedora and RedHat (so CentOS as well) put the mailman data and binaries
in different locations than other systems (looks like the provider
have been mainly for debian so far). For the former os we set
explicit paths for the different used binaries, so this provider works
now as well on those.